+++ hal/src/main/native/main.cpp
@@ -1,0 +1,6 @@
+#include "mau/HALInitializer.h"
+
+int main() {
+    hal::init::InitializeHAL();
+    return 0;
+}
