+++ hal/src/main/native/mau/DriveStation/include/DriverStationInternal.h
@@ -1,0 +1,11 @@
+#ifndef NATIVE_DRIVERSTATIONINTERNAL_H
+
+#include "MauDriveData.h"
+
+#define NATIVE_DRIVERSTATIONINTERNAL_H
+
+namespace mau {
+    extern Mau_DriveData* sharedMemory;
+}
+
+#endif //NATIVE_DRIVERSTATIONINTERNAL_H
